Add the ability to specify dependencies to include proto files from
This will allow plugin users to import proto files from other dependencies in their local proto files. We could search all their deps, but this commit allows users to whitelist specific deps. This was mostly a performance concern.

The proto dependency files will not be compiled, but will be included during protoc invocation.

After adding this support, I did a little refactoring to use the same machinery for finding the standard google proto files. There is no advantage to this way, but it keeps us from having to fix bugs in two places. I still left the addition of the builtin protos on their own b/c the error handling is naturally different.